+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Not released
 
-- Add creatable functionality to Autocomplete [#828](https://github.com/CartoDB/carto-react/pull/828)
+- Add creatable functionality to Autocomplete & MenuItem fixed [#828](https://github.com/CartoDB/carto-react/pull/828)
 - Table component: Added selected row and with checkbox example in Storybook [#876](https://github.com/CartoDB/carto-react/pull/876)
 
 ## 3.0.0